Add debug only ThreadFactorySingleton.
- Add a one-off ThreadFactorySingleton to globally access the ThreadFactory instance corresponding
to the platform.
- Revert the previous attempt using an extern'ed global.
